requirements.txt
asgiref==3.4.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==2.2.16
django-cors-headers==3.4.0
django-filter==2.3.0
django-python3-ldap==0.12.1
django-redis==4.12.1
django-simpleui==2020.9
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
gunicorn==20.1.0
httpsig==1.3.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonfield==3.1.0
ldap3==2.9.1
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.2.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.11.0
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.6
pytz==2021.3
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
